move everything to lib
