+++ aladdin/commands/cluster_init.py
@@ -1,7 +1,7 @@
-from aladdin.cluster_rules import cluster_rules
+from aladdin.lib.cluster_rules import cluster_rules
 from aladdin.commands import deploy
 from aladdin.lib.k8s.helm import Helm
-from aladdin.arg_tools import container_command
+from aladdin.lib.arg_tools import container_command
 
 
 def parse_args(sub_parser):
