+++ aladdin/commands/cmd.py
@@ -3,7 +3,7 @@ import logging
 import subprocess
 import sys
 
-from aladdin.arg_tools import add_namespace_argument, container_command
+from aladdin.lib.arg_tools import add_namespace_argument, container_command
 from aladdin.lib.k8s.kubernetes import Kubernetes
