+++ aladdin/commands/namespace_init.py
@@ -1,5 +1,5 @@
-from aladdin.arg_tools import add_namespace_argument, container_command
-from aladdin.cluster_rules import cluster_rules
+from aladdin.lib.arg_tools import add_namespace_argument, container_command
+from aladdin.lib.cluster_rules import cluster_rules
 from aladdin.commands import deploy
 from aladdin.lib.k8s.helm import Helm
