+++ aladdin/commands/rollback.py
@@ -1,7 +1,7 @@
-from aladdin.arg_tools import add_namespace_argument, container_command
-from aladdin.cluster_rules import cluster_rules
+from aladdin.lib.arg_tools import add_namespace_argument, container_command
+from aladdin.lib.cluster_rules import cluster_rules
 from aladdin.commands import sync_ingress, sync_dns
-from aladdin.helm_rules import HelmRules
+from aladdin.lib.helm_rules import HelmRules
 from aladdin.lib.k8s.helm import Helm
